services/order-service/requirements.txt
# services/order-service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
asyncpg==0.28.0
pulsar-client==3.3.0
pydantic==2.4.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
coverage==7.3.2
httpx==0.25.0
pytest-cov==4.1.0
asyncio==3.4.3
python-jose==3.3.0
python-multipart==0.0.6
